+++ packages/pocket/src/load-more/index.ts
@@ -1,0 +1,5 @@
+import LoadMore from './LoadMore';
+
+export default LoadMore;
+
+export * from './LoadMore';