+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orca-fe/hooks
 
+## 1.6.0
+
+### Minor Changes
+
+- feat: useLoadMore & LoadMore
+
 ## 1.5.1
 
 ### Patch Changes